:art: refactor : 하나의 클래스 내에서만 사용되는 유효성 검증 로직은 따로 Validator 로 빼지 않고 관련 클래스에 넣기
