+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,5 +1,6 @@
 package racingcar.domain;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -42,5 +43,28 @@ class CarTest {
     void makeCars(String name) {
         assertThatIllegalArgumentException()
                 .isThrownBy(() -> new Car(name));
+    }
+    
+    @ParameterizedTest(name = "carName : {0}")
+    @NullAndEmptySource
+    @DisplayName("자동차 이름 입력이 비어있는 경우 예외처리")
+    void validateNullAndEmpty(String carName) {
+        Assertions.assertThatIllegalArgumentException()
+                .isThrownBy(() -> new Car(carName));
+    }
+    
+    @Test
+    @DisplayName("자동차 이름의 길이 범위 초과 시 예외처리")
+    void validateOutOfLength() {
+        Assertions.assertThatIllegalArgumentException()
+                .isThrownBy(() -> new Car("bbbbbb"));
+    }
+    
+    @ParameterizedTest(name = "carName : {0}")
+    @ValueSource(strings = {"aa, b,cc", "aa,b ,cc"})
+    @DisplayName("이름에는 공백을 입력 시 예외처리")
+    void validateOutOfRange(String carName) {
+        Assertions.assertThatIllegalArgumentException()
+                .isThrownBy(() -> new Car(carName));
     }
 }
